tables

+++ migrations/20241128175951_create_accounts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS accounts (
   id TEXT PRIMARY KEY,
-  chat_id INTEGER NOT NULL,
+  chat_id BIGINT NOT NULL,
   first_name VARCHAR NOT NULL,
   last_name VARCHAR NOT NULL,
   verified BOOLEAN NOT NULL DEFAULT FALSE,
